requirements.txt
torchvision
torch
diffusers
pillow
open_clip_torch
clip
tqdm
transformers
scikit-image
datasets
